# docs/swagger.yaml
basePath: /api/v1
definitions:
  common.RESTBody-any:
    properties:
      data: {}
      error:
        $ref: '#/definitions/common.RESTBodyError'
      message:
        type: string
    type: object
  common.RESTBodyError:
    properties:
      code:
        type: integer
      reason:
        type: string
    type: object
  user.SignUpDTO:
    properties:
      email:
        type: string
      fullname:
        type: string
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: wahyurudiyan@gmail.com
    name: Go Boilerplate API Support
  description: This is a sample boilerplate project for golang backend service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Golang Boilerplate Example API
  version: "0.1"
paths:
  /health-check:
    get:
      consumes:
      - '*/*'
      description: the health check endpoint provide the status of server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
      summary: Show the status of server.
      tags:
      - Health Check Endpoint
  /users/signup:
    post:
      consumes:
      - '*/*'
      description: endpoint that handle user register.
      parameters:
      - description: Request body type
        in: header
        name: Authorization
        required: true
        type: string
      - description: Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.SignUpDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/common.RESTBody-any'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/common.RESTBody-any'
      summary: SignUp user endpoint.
      tags:
      - User Endpoint
swagger: "2.0"